Add new file thresholds_common.cpp.
