+++ python/boomer/common/cpp/thresholds.cpp
@@ -1,257 +1,5 @@
 #include "thresholds.h"
-#include <cstdlib>
 
-
-/**
- * A wrapper for a (filtered) feature vector that is stored in the cache. The field `numConditions` specifies how many
- * conditions the rule contained when the array was updated for the last time. It may be used to check if the array is
- * still valid or must be updated.
- */
-struct CacheEntry {
-    CacheEntry() : numConditions(0) { };
-    std::unique_ptr<FeatureVector> featureVectorPtr;
-    uint32 numConditions;
-};
-
-/**
- * Updates the given statistics by applying the weights of the individual training examples.
- *
- * @param statistics    A reference to an object of type `AbstractStatistics` to be updated
- * @param weights       A reference to an object of type `IWeightVector` that provides access to the weights of the
- *                      individual training examples
- */
-static inline void updateSampledStatistics(AbstractStatistics& statistics, const IWeightVector& weights) {
-    uint32 numExamples = statistics.getNumRows();
-    statistics.resetSampledStatistics();
-
-    for (uint32 r = 0; r < numExamples; r++) {
-        uint32 weight = weights.getValue(r);
-        statistics.addSampledStatistic(r, weight);
-    }
-}
-
-/**
- * Adjusts the position that separates the examples that are covered by a condition from the ones that are not covered,
- * with respect to those examples that are not contained in the current sub-sample. This requires to look back a certain
- * number of examples to see if they satisfy the new condition or not. I.e., to traverse the examples in ascending or
- * descending order, depending on whether `conditionEnd` is smaller than `conditionPrevious` or vice versa, until the
- * next example that is contained in the current sub-sampling is encountered.
- *
- * @param featureVector     A reference to an object of type `FeatureVector` that stores the indices and feature values
- *                          of the training examples
- * @param conditionEnd      The position that separates the covered from the uncovered examples when only taking into
- *                          account the examples that are contained in the current sub-sample
- * @param conditionPrevious The position to stop at (exclusive)
- * @param threshold         The threshold of the condition
- * @return                  The adjusted position that separates the covered from the uncovered examples with respect to
- *                          the examples that are not contained in the current sub-sample
- */
-static inline intp adjustSplit(FeatureVector& featureVector, intp conditionEnd, intp conditionPrevious,
-                               float32 threshold) {
-    FeatureVector::const_iterator iterator = featureVector.cbegin();
-    intp adjustedPosition = conditionEnd;
-    bool ascending = conditionEnd < conditionPrevious;
-    intp direction = ascending ? 1 : -1;
-    intp start = conditionEnd + direction;
-    uint32 numSteps = abs(start - conditionPrevious);
-
-    // Traverse the examples in ascending (or descending) order until we encounter an example that is contained in the
-    // current sub-sample...
-    for (uint32 i = 0; i < numSteps; i++) {
-        // Check if the current position should be adjusted, or not. This is the case, if the feature value of the
-        // current example is smaller than or equal to the given `threshold` (or greater than the `threshold`, if we
-        // traverse in descending direction)
-        uint32 r = start + (i * direction);
-        float32 featureValue = iterator[r].value;
-        bool adjust = ascending ? featureValue <= threshold : featureValue > threshold;
-
-        if (adjust) {
-            // Update the adjusted position and continue...
-            adjustedPosition = r;
-        } else {
-            // If we have found the first example that is separated from the example at the position we started at, we
-            // are done...
-            break;
-        }
-    }
-
-    return adjustedPosition;
-}
-
-/**
- * Filters a feature vector that contains the indices of the examples that are covered by the previous rule, as well as
- * their values for a certain feature, after a new condition that corresponds to said feature has been added, such that
- * the filtered vector does only contain the indices and feature values of the examples that are covered by the new
- * rule. The filtered vector is stored in a given struct of type `CacheEntry` and the given statistics are updated
- * accordingly.
- *
- * @param cacheEntry            A reference to a struct of type `CacheEntry` that should be used to store the filtered
- *                              feature vector
- * @param featureVector         A reference to an object of type `FeatureVector` that should be filtered
- * @param conditionStart        The element in `featureVector` that corresponds to the first example (inclusive)
- *                              included in the `IStatisticsSubset` that is covered by the new condition
- * @param conditionEnd          The element in `featureVector` that corresponds to the last example (exclusive)
- * @param conditionComparator   The type of the operator that is used by the new condition
- * @param covered               True, if the examples in range [conditionStart, conditionEnd) are covered by the new
- *                              condition and the remaining ones are not, false, if the examples in said range are not
- *                              covered and the remaining ones are
- * @param numConditions         The total number of conditions in the rule's body (including the new one)
- * @param coveredExamplesMask   An array of type `uint32`, shape `(num_examples)` that is used to keep track of the
- *                              indices of the examples that are covered by the previous rule. It will be updated by
- *                              this function
- * @param coveredExamplesTarget The value that is used to mark those elements in `coveredExamplesMask` that are covered
- *                              by the previous rule
- * @param statistics            A reference to an object of type `AbstractStatistics` to be notified about the examples
- *                              that must be considered when searching for the next refinement, i.e., the examples that
- *                              are covered by the new rule
- * @param weights               A reference to an an object of type `IWeightVector` that provides access to the weights
- *                              of the training examples
- * @return                      The value that is used to mark those elements in the updated `coveredExamplesMask` that
- *                              are covered by the new rule
- */
-static inline uint32 filterCurrentFeatureVector(CacheEntry& cacheEntry, FeatureVector& featureVector,
-                                                intp conditionStart, intp conditionEnd, Comparator conditionComparator,
-                                                bool covered, uint32 numConditions, uint32* coveredExamplesMask,
-                                                uint32 coveredExamplesTarget, AbstractStatistics& statistics,
-                                                const IWeightVector& weights) {
-    uint32 numTotalElements = featureVector.getNumElements();
-    FeatureVector::const_iterator iterator = featureVector.cbegin();
-    bool descending = conditionEnd < conditionStart;
-    uint32 updatedTarget;
-
-    // Determine the number of elements in the filtered vector...
-    uint32 numConditionSteps = abs(conditionStart - conditionEnd);
-    uint32 numElements = covered ? numConditionSteps :
-        (numTotalElements > numConditionSteps ? numTotalElements - numConditionSteps : 0);
-
-    // Create a new vector that will contain the filtered elements...
-    std::unique_ptr<FeatureVector> filteredVectorPtr = std::make_unique<FeatureVector>(numElements);
-    FeatureVector::iterator filteredIterator = filteredVectorPtr->begin();
-
-    intp direction;
-    uint32 i;
-
-    if (descending) {
-        direction = -1;
-        i = numElements - 1;
-    } else {
-        direction = 1;
-        i = 0;
-    }
-
-    if (covered) {
-        updatedTarget = numConditions;
-        statistics.resetCoveredStatistics();
-
-        // Retain the indices at positions [conditionStart, conditionEnd) and set the corresponding values in
-        // `coveredExamplesMasK` to `numConditions`, which marks them as covered (because
-        // `updatedTarget == numConditions`)...
-        for (uint32 j = 0; j < numConditionSteps; j++) {
-            uint32 r = conditionStart + (j * direction);
-            uint32 index = iterator[r].index;
-            coveredExamplesMask[index] = numConditions;
-            filteredIterator[i].index = index;
-            filteredIterator[i].value = iterator[r].value;
-            uint32 weight = weights.getValue(index);
-            statistics.updateCoveredStatistic(index, weight, false);
-            i += direction;
-        }
-    } else {
-        updatedTarget = coveredExamplesTarget;
-        intp start, end;
-
-        if (descending) {
-            start = numTotalElements - 1;
-            end = -1;
-        } else {
-            start = 0;
-            end = numTotalElements;
-        }
-
-        if (conditionComparator == NEQ) {
-            // Retain the indices at positions [start, conditionStart), while leaving the corresponding values in
-            // `coveredExamplesMask` untouched, such that all previously covered examples in said range are still marked
-            // as covered, while previously uncovered examples are still marked as uncovered...
-            uint32 numSteps = abs(start - conditionStart);
-
-            for (uint32 j = 0; j < numSteps; j++) {
-                uint32 r = start + (j * direction);
-                filteredIterator[i].index = iterator[r].index;
-                filteredIterator[i].value = iterator[r].value;
-                i += direction;
-            }
-        }
-
-        // Discard the indices at positions [conditionStart, conditionEnd) and set the corresponding values in
-        // `coveredExamplesMask` to `numConditions`, which marks them as uncovered (because
-        // `updatedTarget != numConditions`)...
-        for (uint32 j = 0; j < numConditionSteps; j++) {
-            uint32 r = conditionStart + (j * direction);
-            uint32 index = iterator[r].index;
-            coveredExamplesMask[index] = numConditions;
-            uint32 weight = weights.getValue(index);
-            statistics.updateCoveredStatistic(index, weight, true);
-        }
-
-        // Retain the indices at positions [conditionEnd, end), while leaving the corresponding values in
-        // `coveredExamplesMask` untouched, such that all previously covered examples in said range are still marked as
-        // covered, while previously uncovered examples are still marked as uncovered...
-        uint32 numSteps = abs(conditionEnd - end);
-
-        for (uint32 j = 0; j < numSteps; j++) {
-            uint32 r = conditionEnd + (j * direction);
-            filteredIterator[i].index = iterator[r].index;
-            filteredIterator[i].value = iterator[r].value;
-            i += direction;
-        }
-    }
-
-    cacheEntry.featureVectorPtr = std::move(filteredVectorPtr);
-    cacheEntry.numConditions = numConditions;
-    return updatedTarget;
-}
-
-/**
- * Filters a feature vector that contains the indices of training examples, as well as their values for a certain
- * feature, such that the filtered vector does only contain the indices and feature values of those examples that are
- * covered by the current rule. The filtered vector is stored in a given struct of type `CacheEntry`.
- *
- * @param indexedArray          A reference to an object of type `FeatureVector` that should be filtered
- * @param cacheEntry            A reference to a struct of type `CacheEntry` that should be used to store the filtered
- *                              vector
- * @param numConditions         The total number of conditions in the current rule's body
- * @param coveredExamplesMask   An array of type `uint32`, shape `(num_examples)`, that is used to keep track of the
- *                              indices of the examples that are covered by the current rule
- * @param coveredExamplesTarget The value that is used to mark those elements in `coveredExamplesMask` that are covered
- *                              by the current rule
- */
-static inline void filterAnyFeatureVector(FeatureVector& featureVector, CacheEntry& cacheEntry, uint32 numConditions,
-                                          const uint32* coveredExamplesMask, uint32 coveredExamplesTarget) {
-    uint32 maxElements = featureVector.getNumElements();
-    FeatureVector* filteredVector = cacheEntry.featureVectorPtr.get();
-
-    if (filteredVector == nullptr) {
-        cacheEntry.featureVectorPtr = std::move(std::make_unique<FeatureVector>(maxElements));
-        filteredVector = cacheEntry.featureVectorPtr.get();
-    }
-
-    FeatureVector::const_iterator iterator = featureVector.cbegin();
-    FeatureVector::iterator filteredIterator = filteredVector->begin();
-    uint32 i = 0;
-
-    for (uint32 r = 0; r < maxElements; r++) {
-        uint32 index = iterator[r].index;
-
-        if (coveredExamplesMask[index] == coveredExamplesTarget) {
-            filteredIterator[i].index = index;
-            filteredIterator[i].value = iterator[r].value;
-            i++;
-        }
-    }
-
-    filteredVector->setNumElements(i);
-    cacheEntry.numConditions = numConditions;
-}
 
 AbstractThresholds::AbstractThresholds(std::shared_ptr<IFeatureMatrix> featureMatrixPtr,
                                        std::shared_ptr<INominalFeatureVector> nominalFeatureVectorPtr,
