+++ python/boomer/common/cpp/thresholds_approximate.cpp
@@ -1,4 +1,5 @@
 #include "thresholds_approximate.h"
+#include "thresholds_common.cpp"
 #include "rule_refinement_approximate.cpp"
