+++ python/boomer/common/cpp/thresholds_exact.cpp
@@ -1,4 +1,5 @@
 #include "thresholds_exact.h"
+#include "thresholds_common.cpp"
 #include "rule_refinement_exact.cpp"
